api_transcrib/requirements.txt
aiohttp==3.8.6
pydantic==2.6.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.1
redis==4.4.2
backoff==2.2.1
orjson==3.9.12
requests==2.31.0
uvicorn==0.27.1
fastapi==0.109.1
psycopg2==2.9.9
starlette==0.35.1
gunicorn==21.2.0
greenlet==3.0.3
sqlalchemy==2.0.27
werkzeug==3.0.1
asyncpg==0.29.0
PyJWT==2.8.0
pydantic-settings==2.2.1
email-validator==2.1.0.post1
alembic==1.13.1
click==8.1.7
typer[all]==0.9.0
kafka-python==2.0.2
marshmallow==3.21.1
confluent-kafka==2.3.0
requests==2.31.0